Add VVVVVV executable detecting function for Linux
So I've had this unstaged for a good while... I mentioned before on
Discord that it would be really nice if Ved was able to detect the path
to VVVVVV's executable by itself - based on just having the game
running - instead of having to manually find the executable on your
disk.

On macOS this turns out to be bizarrely easy (I tried it on macOS
first because the person who initially selected the wrong executable,
before I got this idea, was on macOS). On Windows it turns out to be
realistic, I think? (You can choose between one crazy API or another
crazy API). And on Linux, I guess it still feels kinda cursed and hacky
to me that the way to interact with the system isn't always by calling
C functions, but sometimes it is by starting processes by specifying
their name and arguments in a string, or you have to access a file
(notably the /proc filesystem), and then interpret data textually...
So I programmed it, and with it written down it doesn't seem too bad,
I guess.

So, since I was planning to recompile the macOS libraries this version
anyway (to make them universal Intel+ARM binaries), I figured this was
a good time to increase the version number for Linux as well in
parallel and take the opportunity to prepare for this VVVVVV-detecting
feature. I'm not sure yet whether I'll add it this version, but it'll
need specific code and testing for each OS anyway so doing it in steps
is fine. The Linux function works well now.

This increments the version number of the filefunc library to 04, but
the function is not yet used, so I'll change the number sometime later
this version so it doesn't have to be changed multiple times.

[pre06]