Make VVVVVV detection functional on Linux, compile universal Mac dylibs
The UI for autodetecting VVVVVV is now done, so if you're on Linux, you
can now actually try autodetecting VVVVVV! This officially bumps the
filefunc library version number to 04.

I also took this moment to, in parallel, up the version to 04 on the
Mac side as well, to make that binary universal (Intel+ARM). And then
why not also recompile the Mac HTTPS lib (which is now version 01).

Next up I'll probably add Windows code for VVVVVV autodetection - I
already have it written in C so I'd just need to port it to Lua (and
make it not just spawn debug message boxes)

[pre08]