Add Windows support for autodetecting VVVVVV (almost)
It seems like the Windows Steam version of VVVVVV is different somehow
and can't be detected yet... Probably because I removed some module
iteration code which had no effect when I tested it and which I saw no
need for because you could just pass NULL to the other function and
it'd just work. I really hope that's the sole reason it doesn't work...
VVVVVV launched from Steam on Windows is probably the most important to
be detectable of all editions.

[pre09]