Do case-insensitive check in Windows VVVVVV executable detection
VVVVVV in Steam on Windows, for some reason, is launched with the
all-lowercase base name "vvvvvv.exe". Which is why it couldn't be
found. I'm so glad it was that simple.

Now autodetection should be fully ready on Windows! :D

(this commit also includes the EnumProcessModulesEx declaration and
some constants, which don't really hurt to be there and make it easier
to debug this in the future, should people start reporting that it
doesn't work for them for some mysterious reasons)

[pre10]