Improve library loading, remove filefunc_lin_fallback.lua
The C filesystem library (a .so) for Linux is now provided with Ved
itself, instead of only being locally compiled. It can still be locally
compiled if the provided library doesn't work - for example, you're on
32-bit Linux despite the Linux version of VVVVVV only being distributed
as 64-bit nowadays, or you _are_ using 64-bit but it's so old that many
other things won't run either (I compiled my binary on Ubuntu 18.04
expecting it not to be compatible with CentOS 7, but turns out it is!)

The reason the library was strictly compiled locally is because I've
always read vague warnings that you can't trust Linux binaries to be
portable, and while a binary might run on your system and similar-
enough systems, there are different (versions of) distros, which have
mysteriously different ABIs and whatnot, where it will just... refuse
to load? crash? produce the most horrifying of undefined behavior?
I also read that the fields in struct dirent differ between "systems".

So, given my uncertainty and unconfidence around this, and given that
I have no control over nor insight into what versions of what distros
they package LÖVE for, compiling the code locally felt like the only
safe way to distribute it. And not everyone has GCC installed either,
so we had another implementation of all the filesystem functions, in
filefunc_lin_fallback.lua. This did things like calling `ls` instead.

Now, I finally researched this scary portability/ABI thing better. I
made a handful of VMs to try things out, and found out there's some
protections like sonames and GLIBC_* versions exactly to avoid any
_scariness_. All dependencies I need to dynamically link against keep
their ABIs stable or at least backwards compatible, and otherwise it
should simply refuse to load the library (which can be detected) before
anything goes wrong. So my .so should work for basically everyone, and
if not, Ved can see it and try GCC, or give the user a specific error
message. This also means we can remove filefunc_lin_fallback, so I did.

There's now a new file, librarian.lua, which is meant for functionality
that "manages" the libraries and makes sure they're ready for use. This
cleans up some repeated code and overly verbose file path arguments.

[pre13]