Simplify usage of font8:getWrap and unify LÖVE 11.4 difference
In LÖVE 11.3 and below, passing an empty string to Font:getWrap will
give you an empty table for the lines. In 11.4, you get one element in
the table. The result is that the "Internal scripting mode is enabled/
disabled" tooltip in the top left of the script editor was unexpectedly
too high in LÖVE 11.4 only, because the box had an empty description.

Now there's a function font8_getWrap, which reverts to the behavior of
giving 0 lines if the string is empty. Furthermore, it also removes all
the LÖVE 0.9/0.10 conversions that were underneath every font8:getWrap.
0.9 returned a number of lines, 0.10 started returning a table of
lines. We always convert to the number of lines since we never need the
contents of the lines themselves when calling Font:getWrap - and even
without 0.9 support, there'd just be `#lines` or `#wrappedtext`
everywhere instead of `lines`, so not sure if that method is better.
With this function we don't need to worry about that anymore, we just
get the number we need.

[pre14]