Rewrite Linux and macOS playtesting code to be many times longer
This is what the previous code for starting VVVVVV and piping a level
file to stdin boiled down to, for Linux and macOS:

    local process = io.popen("'" .. escaped_path .. "' " .. args, "w")
    success, err = process:write(levelcontents)
    theprocess:close()

This pretty much works, and depending on several things, VVVVVV starts
up in about 0-2 seconds or so when you start playtesting.

After some discussion on Discord, Misa and I came up with a way to
hopefully make this loading time near-zero in most cases in VVVVVV 2.4,
by already starting VVVVVV the moment you hit the playtesting button
(or maybe even earlier!) and just letting VVVVVV load its assets and
then block on reading stdin BEFORE creating a window, while the player
selects a start point. Once Ved pipes in the level data, VVVVVV is
already loaded and only has to create the window and load the level.

This would probably require only small changes to the Lua above, by
just delaying the process:write(). However, it would be nice to
automatically detect whether the player is using VVVVVV 2.3 or early
2.4 (which doesn't support it yet) or recent VVVVVV 2.4 (which now
does) instead of just assuming the player is using 2.4 when it hasn't
been officially released yet, or asking the player to choose. 2.4 now
has a -version argument since a46c49c, which earlier versions don't
have, but 2.3 still instantly exits saying it doesn't recognize the
flag, so it can still be distinguished.

This is where I discovered io.popen is not really that powerful, so I
started writing a little bit more flexible code in C to replace it...
Now I ended up with three hundred lines of C code! Ved's C segment has
more than doubled from this! Either way, even though it's much longer,
I think it's still well-structured and provides a reasonable API. It's
also a lot more similar to the Windows code than I would've expected,
so I'll soon implement the new functions in filefunc_linmac.lua (which
are basically wrappers for the C functions) on Windows as well, instead
of the Windows-only run_pipe_process. It should then be pretty easy to
add the version check, distinguish 2.2-, 2.3 and 2.4+ from each other
and optimize playtesting based on the version number, etc. (2.2 will
just be killed after a second or two of idling on -version, which we
can now do and also detect). We can also detect errors much better than
with io.popen (you can't even check that io.popen failed, you have to
see from trying to write to the pipe and getting a broken pipe error)

About broken pipe errors, on macOS specifically, my new implementation
had the same problem that the existing io.popen one also had on macOS:
if the executable for playtesting is invalid (for example because the
user chose the .app folder instead of the executable, which Ved's new
VVVVVV autodetection should make much less likely) Ved would just close
when you tried to playtest. Turns out if you attempt to write to a pipe
with no readers, you get a SIGPIPE which, by default, just terminates
your app. No idea why I only saw this happen on macOS (and why I only
saw VVVVVV intermittently get a SIGPIPE on exit on macOS as well from
me closing the stdout read end) but we can now also ignore SIGPIPE.

While I was modifying the filefunc library anyway, I added a ved_mkdir
function for a feature I'm thinking of adding (being able to create
level assets folders) so I don't have to compile again for that later.

[pre15]