Change Windows playtest code to provide same functions as Linux/macOS
This is kinda broken though. While Windows and Linux/macOS are similar
enough that they use the same principle of having pipes, starting a
child process and waiting until it finishes, they have different enough
opinions on what the exact order of operations should be, when which
handle should/can be closed, and when which function should block, that
I can't get it to do what I want without gross changes to the "filefunc
unified Lua" API. For example, I want to wait until a process finishes,
and only then read from its stdout pipe. Unfortunately, it looks like a
process handle isn't signaled as terminated until you close your
reading end of the stdout and stderr handles! Besides, if the pipe
buffer fills up and nobody's reading it (because you want to wait until
the process terminates so you can get *all* its output), apparently the
child process will block on writing anyway! So instead of waiting for
the process to finish, you're supposed to just stall on reading from
the pipe until it fails with a broken pipe error, to know when the
process terminates. So basically, await_process would also need to get
a stdout handle, and then read_from_pipe needs to read from a buffer
you've been storing data in instead of reading from the pipe when you
need it. Maybe it'd be better to not get the stdout pipe if you don't
need it... But have I mentioned that on Linux/macOS, writing to a pipe
that has no open read handles causes the process to be terminated? In
other words, VVVVVV in this case? Even though VVVVVV apparently doesn't
until it exits by itself, or something. And only sometimes. computers.

To make playtesting on Windows not get stuck on "You are currently
playtesting" I'm just not retaining the stdout and stderr handles on
Windows for now, but after this I'm going to rewrite the Lua API to be
more abstract and to allow it to optimize for OS-specific behavior. The
API is kinda not the best anyway...

[pre16]