Make reading stdout/stderr from VVVVVV work well on Windows
It only took:
- ReadFileEx instead of ReadFile (not even regular ReadFile with
  OVERLAPPED!) which means completion callbacks
- Named pipes instead of normal CreatePipe anonymous pipes (wow)
- Event handles to be signalled
- WaitForMultipleObjectsEx instead of WaitForSingleObject
- Lots of confusion from ambiguous wording and incompletenesses in
  Win32 documentation
- Hours of work spread over multiple days

But now I got a robust implementation of cProcess on Windows! And even
with all the debugging and iteration having to choose alternative
designs because the functions I thought would work well had some sort
of caveat, I'm pretty proud of the result lol

The best part of using ReadFileEx (with a completion routine), instead
of ReadFile when overlapped, is that I can call my completion routine
myself when ReadFileEx decides not to - guaranteeing that the handling
of the result can be in ONE place, regardless of whether that result is
delayed, completed immediately, or is an error like "broken pipe".
GetOverlappedResult can't deal with failures that were known
immediately. And it seems clearer when ReadFileEx will call your
completion routine and when it won't, than when ReadFile will signal
your event and make GetOverlappedResult do something useful, anyway.

[pre18]