Add syntax color for mismatched simplified/internal scripting mode
Sometimes, people (including me) want to write an internal script, but
forget to turn on internal scripting mode. The only indicator that you
forgot to turn on internal scripting mode was that the button on the
right was _not_ yellow, which is pretty easy to miss. Recently, in
1.9.1, I added an icon in the top left of the script editor which
changes shape and color to indicate whether internal scripting mode is
on or not, but this is also probably something that starts to blend in.

So here's a new feature that should really help: the script editor
knows which commands are simplified and which are internal (of course
commands can be both), and if you use an internal command while
internal scripting mode is off, or vice versa, that command appears
dark red. This is a new configurable syntax color, so if you don't like
it, you can choose to make it less noticeable, or change it into the
regular command color entirely for example.

I don't really know why I didn't implement this earlier, because I did
think of it before. Maybe I had concerns about manual internal
scripting causing false-positives or something. But it's just a
different color, and it can be turned down/off, so it should definitely
do more good than harm.

[pre19]