Implement VVVVVV version check in playtesting thread
This brings us much closer to supporting "preloaded" playtesting, since
Ved will have to know whether the VVVVVV executable supports it or not.
We can now also distinguish 2.2- and warn the user about it. (Though,
of course, the timeout turns out to be broken now on Linux and probably
macOS! But I already discovered that and after a long time of trying to
debug this (and finding that VVVVVV spawned by Ved didn't even respond
to SIGTERM, or even SIGALRM called by its own code), I found a section
on the signal(7) manpage about blocked signals and signal masks. Yeah
that makes sense I suppose. Guess that's what happens when a secondary
thread continues as the main thread. So I'll need to fix that as well.

Some other small changes:
- Fixed -playassets option, and the fact that the last option was
  ignored on Linux and macOS
- Exit codes above 0x40000000 on Windows are now seen as failure (for
  example, C0000005 means segfault/access violation)

[pre20]