Add full support for 2.4 "preloaded" playtesting
VVVVVV 2.4 is now started when you start playtesting (before choosing
a starting position), and only has to show its window when you do
choose the starting position!

This required a little restructuring, but it's relatively smooth
sailing now due to the earlier changes. playtesting_askwherestart and
playtesting_active have been replaced by a "UI state" (OFF, ASKING or
PLAYTESTING), and there's now also an "engine state" to track the state
of the thread (OFF, IDLE, PLAYTESTING or CANCELING). This is described
more at the top of playtesting.lua.

Funnily enough, the playtesting thread had an incoming Channel which
was completely unused. Now it's used to pass commands, including
different parts of data (the level XML and the starting position)
coming in at different times, and GO and CANCEL to make an IDLE thread
proceed with playtesting or call it off.

[pre21]