Unblock signals for spawned VVVVVV process on Linux and macOS
Good news, bad news, and good enough news.

The good news is that signal blocking in non-main threads is what was
stopping the SIGALRM from working in the forked non-main thread that
was to become a main thread. I now used sigprocmask to clear the
blocked signals (just like how main threads are configured) and it
makes the spawned process respond to signals again.

The bad news? We were exactly using this timeout signal to terminate
VVVVVV 2.2 (and older, or something like that). But 2.2 still doesn't
respond to it! Why not? Because it's launched via that shellscript!
VVVVVV 2.3 behaves the same if you launch it via a similar shellscript.
No idea _exactly_ why, it's probably because the shellscript launches
VVVVVV as _another_ process, and because the signal is sent to bash
instead of directly to VVVVVV 2.3, who knows what will happen! Maybe
bash doesn't care about the signal, but who says the user uses bash,
maybe zsh does something completely different, and maybe fish makes
an actual alarm clock ring or whatever. It doesn't seem like bash is
terminated independently and the VVVVVV it spawned isn't - Ved does
wait until VVVVVV itself is closed.

The good enough news? At least for my bash and macOS zsh, once you
manually close immune-to-the-alarm-VVVVVV 2.2, as long as you do it
after 2 or more seconds (which you probably will because you're first
wondering whether it works or not), Ved will now still see that it
terminated from a SIGALRM, and give the error message. So whatever.
The timeout works on Windows, on Linux and macOS autodetection should
detect 2.2 with a special message anyway, and if you do use 2.2 on
Linux or macOS you'll probably still get a message after closing it.

[pre22]