Change directory_exists(where, what) into directory_exists(path)
The `where, what` arguments were a relic from command-line filesystem
functions (and not just the Linux fallback functions which were removed
recently, but also on Windows and macOS), in that you needed to list
the files in the parent directory and find the matching one. Even if
one of the filefunc implementations were to still need it, it'd make
sense to make the API take a full path and just split it in that
function's implementation.

[pre24]