Improve Assets screen and make it level-specific
This has been kinda needed since 2.3 added support for level-specific
assets, but most of the existing assets tools had been workable
enough... The music editor could load and save files anywhere, the
graphics viewer could load from anywhere, it was only the sounds player
that really didn't work.

Now, there's an Assets button in the main editor as well. This brings
you to the same assets screen as the main button from the loading
screen, but everything points to the assets for that level. So, the
vvvvvvmusic.vvv/mmmmmm.vvv options and the sounds player come to life
again, and any folder open dialogs now start in the level's assets
folder by default.

There are also various other improvements, which apply to viewing the
main VVVVVV folder as well:
- All the options are now placed in a tree-like structure, with the
  assets folder itself represented at the top. You can use this option
  to open the assets folder in your file manager/Explorer/Finder, or
  to *create it if it doesn't exist yet* (actually, that's still a
  todo, but that's easily done)
- The sounds and graphics options now tell you whether their respective
  folders exist, and if they do, how many assets have been modified in
  them (e.g. "1 sound effect customized" or "9 images customized")

[pre24]