Fix crash when jumping to out-of-bounds warp token/start point
gotoroom(rx,ry) (the Ved function) assumed the coordinates given were
valid, and just set the current room to what you told it to. This could
crash jumping to a warp token destination that's out of bounds, or
jumping to a start point entity that's out of bounds. Maybe there's
other cases, but now gotoroom has a check to ensure the requested room
is within map bounds, and gives a dialog and returns false if not.

The code for jumping to warp token destinations/entrances is now also
refactored a bit to make it more readable.

[pre27]