Switch over to using love.wheelmoved
The title is kind of oversimplified - obviously in LÖVE 0.10+ we were
using love.wheelmoved, but until now it was just a wrapper for calling
love.mousepressed with "wu" or "wd" anyway. Now the roles are reversed.

LÖVE 0.9.x has the style of just calling love.mousepressed with the
button constants "wu" or "wd" if the mouse wheel was moved, and
LÖVE 0.10.0 changed it to calling love.wheelmoved with horizontal and
vertical movement values; mostly -1, 0 or 1, but it can also be
higher - but not that often and not that much higher - and nothing says
it can't be non-integer like in SDL 2.0.18+.

As hinted at, love.wheelmoved is now implemented to do all the work,
and on LÖVE 0.9 we call it from love.mousepressed with a constant value
of -1 or 1 if wu/wd buttons were "pressed".

[pre28]