Add support for displaying recolored one-way tiles
The one-way tiles in VVVVVV have always been yellow. This may look bad,
so 2.3 changed it so they would be recolored to the tileset's color,
with BlitSurfaceTinted. This function converts the RGB colors to
grayscale (using a certain formula, because red, green and blue have
different brightnesses so they need to be weighted) and then applies
the tileset color to it. And level-specific tilesets wouldn't be
recolored, unless onewaycol_override is set.

To replicate this recoloring, I made a shader (shader_tint.frac).
I had never made GLSL shaders before, but it turns out they're very
good at this sort of thing (otherwise I would've had to do some kind
of pixel-by-pixel drawing or ImageData:mapPixel and more state-keeping
separate from the state I already have to keep for the SpriteBatch.)

I tested this on a lot of different OSes, OS versions, computers and
LÖVE versions, and it worked fine on all of them (regardless of LÖVE
version even - shaders have always used 0..1 color values) EXCEPT for
my old Windows XP cardboard box, which has a graphics card that doesn't
support shaders. So in that case (or if the shader fails to load for
any other reason) the tiles will just remain yellow.

There is now a second SpriteBatch, tile_batch_oneway, which one-way
tiles are added to instead of to tile_batch. tile_batch_oneway is drawn
separately from tile_batch, and the shader and color are set for it
specifically. One-way tiles actually placed in the room will be
retinted, the cursor and tiles picker won't - but that doesn't really
matter in my opinion, the room itself is drawn accurately. And of
course, it looks at whether or not the tileset image is a level-
specific asset and whether or not onewaycol_override is set.

This ticks one item in #133.

[pre32]