Initialize cVedQueueableSource with Decoder instead of SoundData
If you have a music file with a lot of songs that have loop points,
it would take a while to load, since each looping song was immediately
initialized with a SoundData. And we do need access to the song
duration before playing the song (unless we want an annoying interface
that doesn't display durations for any songs you haven't played yet)

Luckily, you can also create a Decoder, which does very little by
itself (it doesn't decode the sound data until you tell it to) but it
does immediately give us the information we need. Then, when playing
the song, we'll actually initialize the SoundData, and instead of doing
it directly from the file data, we can do it from the Decoder as well.

[pre34]