Update Space Station font
The previous version was from 2018 and was still made in FontStruct, so
it became kinda overdue to update it, but since it's in the
localization branch ready to be included in 2.4, and the final touches
for that seem to be done, now's a pretty good time to update it.

By the way, here's the repo for it that now exists:
https://github.com/Fussmatte/space-station-font

I hope the positioning, line height, etc will be consistent across OSes
and LÖVE versions - the difference in offset between LÖVE 0.10 and 11
that we need to account for reduced from 2 to 1, but somewhat recently
there was another TTF (Indienova Chinese font) that had different
offsets between equivalent LÖVE versions on Windows and Linux for some
reason. I dunno, TTF has too many parameters for a simple monospaced
font like this, and it's annoying that you can't just ignore them,
while different software interprets them differently for some reason.

[pre35]