Force Space Station height to 8 pixels instead of 9
I updated the font in da8970acb74a0a6f11a73cf3b1e8095236a5417d, after
four years of not updating the font. For some reason, the new font has
a line height of 9, not 8 like we used to have. And it's TTF, so there
are a lot of parameters in the name of typography which are all set to
incomprehensible values, and modifying them will not result in the
change you want to see, until you get the change you want to see but
only on Windows, suddenly introducing differences between operating
systems in your font.

For now, I've fixed this with font8:setLineHeight(8/font8:getHeight())
(the default line height is 1.0, so it's a factor). I guess this
function is exactly the way I'm supposed to solve this problem, but I
hope having a font consisting of outlines being rendered at 0.888888889
intervals of what LÖVE thinks its height should be won't cause any
rendering issues somehow, or on some OSes... I'm considering making my
own text renderer for LÖVE for more flexibility.

[pre37]