Make song durations in music player work in LÖVE 0.9.x
The music player was pretty crippled in 0.9.x. Source:getDuration()
didn't exist yet in 0.9.x, so the progress bar couldn't work, and
seeking to anywhere but the start couldn't work either.

I know it's probably not worth it anymore to invest a lot of time and
effort into LÖVE 0.9.x support, but that's the nice part - with the
cVedSource and cVedQueueableSource wrappers that I had to make for loop
points in Ved 1.10.0, it was really easy to fix this issue. Even though
Source:getDuration() didn't exist yet in 0.9.x, SoundData:getDuration()
did (I also didn't realize this earlier), so when playing the song for
the first time, you can just make a temporary SoundData just to get the
duration, and keep that around in an attribute of the wrapper class.

Now the only restrictions in the music player/editor in LÖVE 0.9.x are
that you can't see the duration of songs before playing them for the
first time, and the songs take a little longer to start playing.

[pre38]