Add custom text renderer for replacing LÖVE Font objects
As described in earlier commits, LÖVE's built-in support for fonts
causes certain problems and it's worth replacing it by something else
in Ved. Basically, having to use the stubbornly typographic TTF format
with unpredictable behavior and weird position offsets that also depend
on LÖVE version, and sometimes even OS, has held me back from updating
the stable font since 2018. When I finally did update it, it had a line
spacing that was inexplicably too high. It takes special effort to make
or edit a TTF, and all the values you can tweak are set to inexplicable
values and have devilish behavior when you change them.

So I implemented my own text renderer which can take the same format as
VVVVVV! On a technical level, it stores the font in an Image, has a
table from codepoints to character widths and Quads that point to the
different glyphs, and uses a set of SpriteBatches to place the glyphs
on screen. The SpriteBatches are cached from one frame to the next, so
if print calls don't change, they're just reused. It supports one
fallback font, so a level-specific font.png can be the main font with
the built-in font.png as a fallback (both are combined into one image).

So, Space Station is now completely font.png only. tinynumbersfont
(which I'm renaming tinyfont) is still an ImageFont, but I'm planning
to replace that as well.

Here's all the advantages I can think of in a row:
- Again, TTF is way too typographic, we don't need to deal with ascent,
  baseline, descent, kerning, line height, etc which are all set to
  completely incomprehensible values while the end result is wrong
- Less chance that we get inconsistencies between OSes, systems and
  LÖVE versions. TTF is apparently interpreted differently by LÖVE on
  different OSes. Sometimes there's also bugs in font support that are
  fixed in later LÖVE versions - think of the different vertical
  offsets between LÖVE 0.10 and 11, and I also remember hearing about
  some *horizontal* half-pixel offset bug somewhere in LÖVE 12? As well
  as a bug that Misa reported for LÖVE 12 at some point about lg.printf
  wrapping one character even when the wrap limit was set to the text
  width. And of course, functions are changed by design - Font:getWrap
  works differently between 0.9-0.10 and between 11.3-11.4, for
  example. Now we can have our own wrapping and it's always consistent.
- No conversions between formats anymore - Ved's custom font.png
  support converted font.png into an ImageFont strip and had an entire
  function dedicated to mapping pixels in one image to pixels in
  another image. The built-in font was provided as both a TTF and an
  ImageFont for fallback when font.png was used, which I didn't update
  when I updated the TTF. There's now one font format instead of three.
- More flexibility: With TTF, ImageFont and BMFont, you can't decide
  what happens when a character isn't in the font (unless you're
  FontStruct, heh). Now I could also implement things like my own RTL
  support if I need to (something which LÖVE doesn't support)
- We can even support custom font.png in LÖVE 0.9.x! Ved has never
  supported font.png in 0.9 because ImageFonts weren't customizable
  enough, and now we suddenly do support it. Which means less version-
  specific checks in the code (this now always works), and less
  suffering for those who still use 0.9 for whatever reason.
- I can also talk about ACTUALLY ancient LÖVE versions: 0.7.x didn't
  support UTF-8, so the multi-language "no support" screen was all
  mojibaked. Now that's suddenly fixed too for free. The version-
  specific code to support the text renderer on these old versions is
  minimal, there haven't been that many changes to the API's I needed.
- Speaking about UTF-8 - attempts to print invalid UTF-8 don't have to
  crash Ved entirely anymore. Well, I could of course have put the
  prints inside ved_print[f] inside a pcall, or added some UTF-8
  sanitizing function around it.
- It looks like this is faster/more optimized than LÖVE's text renderer
  in normal circumstances.

[pre39]