Fix potential table reuse bug with cProcess on Windows
I ran into a bug with my own font object (cVedFont) where table
attributes would be shared across all instances of the class. This has
something to do with how metatables work. As far as I can tell this is
the only other place where it has the potential to cause problems, so
I'm just assigning the initial table in the :new method instead of on
the metatable.

[pre40]