Fix map export mis-scaling if Ved is scaled and dialog is kept open
This is a similar problem to the one that was fixed in acd854a (where
the map would be mis-scaled if Ved was scaled and dialog animations
were disabled and you clicked the Save button instead of pressing
Enter). Now, the bug was the same, except instead of dialog animations
being disabled leading to this, it was checking the option "Keep dialog
open".

So, I fixed a bug by moving when the dialog button handler takes place,
but I forgot that this is one of the few dialogs in Ved that can take
an action while the dialog is kept open, which happened to leave
another way for the bug to appear. Because the dialog doesn't close
when it's... set to stay open, the handler is called right after the
validation handler, which is in love.draw if you click the button.

The way I fixed it now is to just add a love.graphics.push() and
love.graphics.origin() before the export, and a love.graphics.pop()
after. That way, whether there's scaling or not, and whether it's in
drawing code or not, any scaling is reset and then restored.

Thanks to Omegaplex for reporting this.

[pre40]