Migrate tinyfont (formerly tinynumbersfont) to VedFont
Now we use VedFont instead of LÖVE Fonts for everything.

The tiny font is very variable width, but this was foreseen: all glyphs
have to have the same width and height in pixels, but the amount of
pixels that the cursor advances can be customized per glyph. (By the
way, I renamed the glyph-specific "width" to "advance", similar to how
BMFont calls it.)

I added an XML file (tinyfont.xml), which I'm also planning to support
in VVVVVV itself for CJK support. This contains the width and height of
all glyphs, as well as a list of codepoint ranges that have a special
advance.

Instead of using fallback fonts for the special cases for macOS and
German and French languages, I added a simple font manipulation method,
cVedFont:copy_glyph. This simply assigns the glyph for one codepoint to
another, so the Cmd, Strg and Echap keys are now in the main font (as
m, ß and é; we ran out of ASCII letters), and they're simply copied
to overwrite Ctrl or Esc. Relatedly, since we're no longer using
Font:setFallbacks, LÖVE 0.9.x now supports macOS- and language-specific
fonts as well.

[pre40]