Add option for enabling/disabling automatic script bumping
It has been requested a few times that an option be added to not move
scripts to the top of the list when you edit them (without having to
right click and selecting the option "without bumping"). So now you can
disable automatic bumping in Ved options! If you disable it, going into
a script won't move it to the top, and the right click options to edit
without bumping change into edit *with* bumping (so you can still bump
scripts if you wish).

[pre41]