Add support for music files having up to 127 songs
This ticks one of the last items in #133, because VVVVVV 2.3 added
support for bigger music files. It also changed the way it loads them
slightly; notably, the first header marked invalid no longer terminates
the headers list. This is all accounted for in Ved now, as well as
loading music files with songs out of order. The interface now has a
page chooser where you can choose which page to view. Pages beyond the
last song are darkened, or not selectable at all if you're just in the
viewer.

VVVVVV 2.3 allows for 128 songs in the music file, but because of the
metadata format, we only allow 127 songs for now - at least until we
want a special editing mode that strips all metadata and only then
allows the last song to be used. The last song is practically useless
unless 127 songs really isn't enough for you (we used to have 16!) or
you have a desire to address song number 127 specifically, when I can't
see a reason why. Note that including songs in non-sequential slots is
already kind of a hack - VVVVVV changes the song numbers to be
sequential when it loads them, so you have to mark the headers for all
songs in between as valid, but have their data be a single byte, so
you'd need to specifically waste slots to get to the last one.

[pre42]