Scale crash screen "background" properly
The crash screen is transparent and shows the last frame before the
crash (or at least, as close as we can get to it). In cd8fa3d, I made
the crash screen itself scale properly according to the scale setting,
but I forgot that the background screenshot is already scaled too, so
it would be double-scaled. For example, in 2x scale, you'd see an
enlarged top left corner of the last screen in the background, behind
an otherwise properly scaled crash screen.

The solution is to uncomment an `s.pscale^-1` part for the
love.graphics.draw call that was apparently already commented in the
initial commit. This time not cleaning up after myself helped save
some time!

[pre43]