Make printing nil values not crash
Most of the time when something like ved_print(nil) or
font8:getWidth(nil) is called, it's because a string is missing from a
language file. This is not really something worth crashing for. And
since we now use VedFont, I can decide to make it simply print "nil"
instead.

[pre43]