Add scripting context for unrecognized/wrong-language commands
Completely unrecognized commands have always shown up red, and since
about a month ago, there's also a warning color for commands that are
recognized, but are from internal scripting while internal scripting
mode is disabled, or vice versa. However, different colors may not be
immediately obvious to someone who hasn't checked the syntax
highlighting color settings. Therefore, with this change, if you're on
a line with an unrecognized or wrong-language command, the syntax
highlighting description will be given in the right bar (in the same
place as where other command-specific information is shown). This is
shown in the same color as the command itself.

Forgetting to turn on internal scripting mode has really happened to
the best of us! The dark red color alone would stop it from happening
to anyone aware of it (or anyone wondering what they did wrong that
everything they type shows up in an "error" color), and now the added
text should make sure mistyping a command or forgetting to switch modes
never happens to anyone again!

Video: https://tolp.nl/host/2022-07-20_14-35-31.mp4

Another small change: the wrong-language color now also shows up if you
may not have finished typing the command, but the command already fully
matches a command in the other language. You don't want a command that
you're still typing and thus obviously doesn't exist, to immediately
become red and highlighted as an invalid line (sometimes called "Java
IDE behavior"), so it doesn't. But for wrong-language commands, this
can only happen if the partial command _is_ already specifically
recognized as a valid command.

[pre45]