Sync language files
This includes translations for all new 1.10.0 strings, as well as
changes to various other languages, particularly Esperanto and
Russian.

For Dutch, I decided to get rid of all ĳ ligatures and replace them all
with normal ij, since the ligature is no longer the lesser of two evils
with the new font. (The way I've always seen it: even if this is a
thing that most people won't really notice, the i and j used to be *so*
misaligned and far spaced apart, that most people would think the
ligature looked better than the separate letters, even with how oddly
*close* together the ligature is. Who knows, maybe I was wrong about
that, and the weird ligature actually catches a lot more people's
attention than separate letters would. But glad that dilemma is gone)

[pre45]